pyproject.toml
[project]
name = "azfinsim"
version = "0.0.1"
dependencies = [
    "colorlog",
    "numpy",
    "opencensus-ext-azure",
    "pandas",
    "psutil",
    "redis"
]
